fix: change search state and component names
